Abstract the 2D and 3D canvasses
