+++ src/lib/p5-components/components/Canvas.svelte
@@ -1,171 +1,10 @@
 <script lang="ts">
-  import { page } from '$app/stores';
-  import { activityStore } from '$lib/activityStore';
-  import ActionButtons from '$lib/components/ActionButtons.svelte';
-  import FormulasAndActivityPanel from '$lib/components/FormulasAndActivityPanel.svelte';
-  import ShareWindow from '$lib/components/ShareWindow.svelte';
-  import SliderPanel from '$lib/components/SliderPanel.svelte';
-  import ToggleSliders from '$lib/components/ToggleSliders.svelte';
-  import { Sliders } from '$lib/utils/Slider';
-  import { onMount } from 'svelte';
-  import { Vector2 } from 'three/src/Three';
-  import type { FnToDraw } from './CanvasUtils';
+  import AbstractCanvas from '$lib/components/AbstractCanvas.svelte';
   import P5Canvas from './P5Canvas.svelte';
-
-  export let enablePan = false;
-  export let sliders = new Sliders();
-  export let title = '';
-  export let background = '#ffffff';
-  export let zoom = 29;
-  export let cameraPosition = new Vector2(0, 0);
-  export let showFormulasDefault = false;
-  export let isIframe = false; // Is the scene inside an iframe?
-
-  let isPlayingSliders = false; // Are any of the sliders being changed AUTOMATIC?
-  let isChangingSliders = false; // Are any of the sliders being changed MANUALLY?
-  let isFullscreen = false; // Is the scene fullscreen?
-
-  let showFormulas = showFormulasDefault; // Show the formulas panel (if it exists)
-
-  let resetKey = Math.random();
-  let height = 0;
-  let width = 0;
-
-  let sceneEl: HTMLDivElement;
-  let fnsToDraw: FnToDraw[] = [];
-
-  $: title = $page.url?.searchParams?.get('title') || title;
-
-  /**
-   * Reset camera position, rotation and sliders.
-   */
-  function reset() {
-    sliders = sliders.reset(); // Reset sliders to default values
-    resetKey = Math.random(); // Update the key to reset the set camera component
-
-    fnsToDraw = [];
-  }
-
-  function pause() {
-    reset();
-    activityStore.reset();
-  }
-
-  function waitThenReset() {
-    if (isIframe) {
-      activityStore.disableAfterAnd(60000, reset);
-    }
-  }
-
-  onMount(() => {
-    const params = $page.url?.searchParams;
-
-    if (sliders.fromURL) {
-      sliders = sliders?.fromURL(params?.get('sliders') || '') || sliders;
-    }
-    isIframe = JSON.parse(params?.get('iframe') || 'false') || isIframe;
-
-    if (!isIframe) {
-      activityStore.enable();
-    }
-  });
 </script>
 
-<div class="rounded overflow-hidden h-full drawer" bind:this={sceneEl}>
-  <!-- For screen readers / accesability acces to toggle the drawer  -->
-  <input id="my-drawer" type="checkbox" class="drawer-toggle" />
-
-  <div
-    role="button"
-    tabindex="0"
-    class="canvasWrapper border-l-4 border-gray-400 drawer-content"
-    class:active={$activityStore}
-    class:isIframe
-    bind:clientHeight={height}
-    bind:clientWidth={width}
-    style="height: var(--height, 100%); background: {background}"
-    on:click={activityStore.enable}
-    on:mousedown={activityStore.enable}
-    on:keydown={activityStore.enable}
-    on:mouseenter={activityStore.removeTimeOut}
-    on:mouseleave={waitThenReset}
-  >
-    <!-- P5 SCENE -->
-    {#key resetKey}
-      <P5Canvas size={{ width, height }} bind:fnsToDraw>
-        <slot />
-      </P5Canvas>
-    {/key}
-
-    <!-- TITLE PANEL -->
-    {#if title && (!isIframe || isFullscreen)}
-      <div class="menu absolute left-2 top-2 bg-base-100 rounded-lg p-4">
-        {title}
-      </div>
-    {/if}
-
-    <!-- SLIDER PANEL -->
-    {#if sliders?.sliders?.length > 0}
-      <SliderPanel isInset={!isIframe || isFullscreen}>
-        <ToggleSliders
-          bind:sliders
-          bind:isPlaying={isPlayingSliders}
-          on:startChanging={() => (isChangingSliders = true)}
-          on:stopChanging={() => (isChangingSliders = false)}
-        />
-      </SliderPanel>
-    {/if}
-
-    <!-- FORMULAS AND ACTIVITY PANEL  -->
-    <!-- Only show if there are formulas and (showFormulas is shown OR not an iframe OR is fullscreen) -->
-    <FormulasAndActivityPanel
-      {isIframe}
-      {isFullscreen}
-      {showFormulas}
-      {isChangingSliders}
-      hasFormulas={$$slots.formulas}
-      on:pause={pause}
-    >
-      <slot name="formulas" />
-    </FormulasAndActivityPanel>
-
-    <!-- ACTION BUTTONS -->
-    <ActionButtons
-      {isIframe}
-      {sceneEl}
-      hasFormulas={$$slots.formulas}
-      bind:isFullscreen
-      bind:showFormulas
-      on:reset={reset}
-    />
-  </div>
-
-  <!-- SHARE WINDOW -->
-
-  <ShareWindow {sliders} />
-</div>
-
-<style lang="postcss">
-  :global(html:has(.isIframe)) {
-    background: white;
-  }
-
-  .canvasWrapper {
-    position: relative;
-    width: var(--width, 100vw);
-    height: var(--height, auto);
-    overflow: hidden;
-
-    &.isIframe {
-      @apply border-l-4 border-gray-400;
-
-      &.active {
-        @apply border-blue-500;
-      }
-    }
-  }
-
-  :global(.canvasWrapper > canvas) {
-    position: absolute;
-  }
-</style>
+<AbstractCanvas let:width let:height>
+  <P5Canvas size={{ width, height }}>
+    <slot />
+  </P5Canvas>
+</AbstractCanvas>
