+++ src/lib/p5-components/components/P5Canvas.svelte
@@ -116,6 +116,8 @@
 
   onDestroy(() => {
     draggables.reset();
+
+    fnsToDraw = [];
   });
 </script>
